requirements.txt
apache-airflow-providers-snowflake>=6.4.0
dbt-core
dbt-snowflake